uses 64 bit divisor masks, helpful on input with more than 32 variables and negligible overhead for input with less than/equal to 32 variables

+++ src/primitives.rs
@@ -5,7 +5,7 @@ pub type ExpVec = Vec<Exponent>;
 // Hash table stuff
 pub type HashValue = u32;
 pub type HashTableLength = u32;
-pub type DivisorMask = u32;
+pub type DivisorMask = u64;
 pub type MonomVec = Vec<HashTableLength>;
 
 // Field stuff
